renamed src to more purposeful package name.  pulled main script to top level. adjusted import statements
