--- run.py
@@ -1,6 +1,0 @@
-from src.main import main
-from web.app import app,socketio
-
-if __name__ == "__main__":
-    # socketio.run(app,debug=True)
-    main()
